Improve log viewer (#45190)
Fixes two bugs:

When selecting text in the log viewer, a re-render of the viewer will de-select text. Even if the log viewer was not changed. Now, we will avoid re-rendering the log viewer unless the props have changed.
When selecting text, the line numbers no longer get selected. We moved the line numbers to be pseudo-elements instead of spans.