[core] New method `ray.experimental.get_local_object_locations` to get an object's location and size info without RPC. (#45071)
This method is "best effort" in that it only returns if the core worker
holds reference to that object ref, but it guarantees to not fire any
RPCs so it's fast to call.

Also fixed a bug: when the object size is unknown (e.g. task is still
running), the function `get_object_locations` used to return an
object_size of 2**64-1. Now, it's fixed to return a object_size of None.
(Fixes #16321). The new function `get_local_object_locations` has the
same behavior.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>