Accelerated DAG channel lock fixes (#45119)
This PR fixes two issues:
1. When the error bit is set on a mutable object, a reader may return
from ReadAcquire() without releasing the channel lock. This will cause
deadlock if another reader calls ReadAcquire() on the channel.

2. Abseil reports a deadlock on the channel lock, though this is a false
positive. Note that Abseil only tracks deadlocks when compiled with
debug mode (`export RAY_DEBUG_BUILD=debug`), so this issue will not
appear otherwise.
Specifically, the following occurs:
(1) Abseil sees ReadAcquire() acquire the destructor lock followed by
the _input_ channel lock.
(2) When WriteAcquire() is called on the _output_ channel, Abseil sees
that the thread already holds the _input_ channel lock and then the
thread acquires the destructor lock.
(3) Since Abseil sees these locks acquired in the reverse order in (2),
it thinks there could be a deadlock.

The deadlock check can only be disabled globally, which is undesirable,
rather than for a specific lock. To resolve this issue for now, I
replace absl::Mutex with std::mutex so that Abseil does not track
deadlock on this lock.

A better long-term solution is to rework the synchronization model so
that a lock is not held after ReadAcquire() returns.

---------

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>