Fix sem_unlink() error check in MutableObjectManager (#45286)
This fixes #45061.

The core worker and the raylet each have their own MutableObjectManager
instance, and when both a reader and a writer are on the same machine,
the reader and writer will each register the same object with separate
MutableObjectManager instances. Thus, when the second instance is
destructed, it will call sem_unlink() on the same two semaphores. As the
two semaphores have already been unlinked by the first instance, the
sem_unlink() calls in the second instance will both fail with ENOENT.

Prior to this PR, we checked that *all* calls to sem_unlink() return 0,
which is incorrect.

---------

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>