[doc] add flowdapt integration to doc ecosystem page (#45281)
Following a discussion in the slack channel with @richardliaw, we are
adding Flowdapt to the ecosystem page since it offers a robust workflow
configuring/scheduling/triggering interface to Ray.

Some background on Flowdapt:
Over at Emergent Methods, we depend on
[Flowdapt](https://docs.flowdapt.ai) for production workflow deployment
for our product https://asknews.app. Meanwhile, we have a variety of
other integrations such as Nowcast, which is designed to adaptively
retrain/inference weather forecasting models on thousands of cities
simultaneously. Some of our clients depend on Flowdapt for QAQC for
high-throughput data analysis.

One cool feature of Flowdapt is that we designed it to accommodate a
variety of backends, including Ray and Dask, as well as a local
executor. Swapping between backends requires zero-code changes for the
application - which allowed us to run a big apples-to-apples comparison
of Ray vs Dask: https://docs.flowdapt.ai/examples/rayvsdask/. Spoiler
alert: Ray wins.

Some open-source example plugins available that run on Flowdapt (and are
best run with the Ray backend):

Weather forecasting:
https://github.com/emergentmethods/flowdapt-nowcast-plugin

Cryptocurrency price forecasting:
https://github.com/emergentmethods/flowdapt-cryptocast-plugin





Signed-off-by: robcaulk <rob.caulk@gmail.com>