[core][experimental] Add support for dynamically sized torch.Tensors passed via NCCL in accelerated DAG (#45332)
This adds support for dynamically sized torch.Tensors to be passed
between accelerated DAG nodes via NCCL. Specifically, the following code
is now supported, whereas previously `shape` and `dtype` had to be
explicitly passed to `TorchTensorType`.

```python
    with InputNode() as inp:
        dag = sender.send.bind(inp)
        dag = dag.with_type_hint(TorchTensorType(transport="nccl"))
        dag = receiver.recv.bind(dag)

    compiled_dag = dag.experimental_compile()
```

The feature works by creating a shared memory channel to pass the
metadata for the shape and dtype of the tensor. The metadata is then
used to create a buffer of the correct size on the NCCL receiver.

Initial microbenchmarks shows this adds about 50% throughput overhead
compared to statically declaring the shape and dtype, or about 160us/DAG
call. This seems a bit higher than expected (see also #45319).

This also adds a few other fixes:
- adds support for reusing actors to create new NCCL groups, which is
needed if a DAG is torn down and a new one is created.
- adds a lock to DAG teardown, to prevent the same NCCL group from
getting destructed twice.
- User-defined TorchTensorType shape or dtype is now used as a hint for
the buffer size, instead of a required size. Since buffers are currently
static, an error will be thrown if the user tries to return a too-large
tensor.

Part 1 of #45306, will follow up with a separate PR for nested tensors.


---------

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>
Co-authored-by: SangBin Cho <rkooo567@gmail.com>
Co-authored-by: Kai-Hsun Chen <kaihsun@apache.org>