[core][experimental] Support NCCL-based torch.Tensors nested inside Python objects (#45473)
Allows torch.Tensors nested inside Python objects to be transferred via
NCCL using the following syntax:

```python
    with InputNode() as inp:
        dag = sender.send.bind(inp)
        dag = dag.with_type_hint(TorchTensorType(transport="nccl"))
        dag = receiver.recv.bind(dag)
```

We implement this by using an additional shared memory channel to pass
CPU data, with a "nested" NCCL channel to pass the GPU data. Here is the
send procedure for the above code:
1. Serialize the data. Extract out all tensors that are on the GPU and
replace them with some kind of placeholder.
2. Send a list of metadata through the meta_channel.
3. Send the GPU tensors through the NCCL channel.
4. Send the rest of the CPU data through a cpu_data_channel, with the
placeholders for the GPU tensors.

Note that if the TorchTensorType doesn't have a shape and dtype
specified, we currently use the separate meta_channel to pass metadata
for the serialized tensors, as introduced in #45332. To elide the
cpu_data_channel, the user should now use
`TorchTensorType(direct_return=True)`, to indicate that no CPU data is
sent along with the GPU data. To elide the meta_channel, the user should
declare the shape and dtype, e.g., `TorchTensorType(shape=(10, ),
dtype=torch.float16)`.

## Related issue number

Closes #45306.

---------

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>
Co-authored-by: SangBin Cho <rkooo567@gmail.com>