[Core] Add placeholder callback methods as members of RayDaskCallback (#45160)
## Why are these changes needed?

This PR adds the overridable methods for `RayDaskCallback` to the base
class.

Previously, the user had a few choices for instantiating a
`RayDaskCallback`:

```python
RayDaskCallback(ray_presubmit=my_custom_presubmit_func)
```

or

```python
with RayDaskCallback(ray_presubmit=my_custom_presubmit_func) as cb:
    ...
```

or by subclassing:

```python
class MyCallback(RayDaskCallback):

    def _ray_presubmit(self, task, key, deps):
        ....
```

Currently the constructor for `RayDaskCallback` dynamically generates
the callback methods and attaches them to the class instance upon
instantiation using `setattr`. This pattern doesn't work with developer
tooling including language servers, static code analysis tools, or
documentation generators because these functions are generated
dynamically. Additionally the documentation for dask-on-ray instructs
users to use the third approach (using subclassing) rather than passing
methods to the constructor. Finally, the docstring for
`RayDaskCallback.__init__` currently has function signatures and
pseudo-docstrings for the callback methods the user should be
implementing. This docstring is itself malformed, so Sphinx mistakenly
tries to interpret parts of this as references which do not exist.

At the request of @c21 I've actually implemented these functions as they
appear in the docstring. Type hints are incomplete here, but match what
was already in the previous docstring. I also added a summary in the
dask-on-ray documentation.

Unblocks #39658.