[runtime_env] unify `container` and (new) `image_uri` under plugin mechanism
Currently, we create a custom `ContainerManager` class and invoke it specially with `container_manager.setup()`. This doesn't conform with the runtime env plugin pattern we use for every other runtime env field.

This PR removes `ContainerManager` and moves the container code to `ContainerPlugin` which extends the base `RuntimeEnvPlugin` class. It also introduces the `ImageURIPlugin` which is the same implementation, but uses a different, new API.

`ContainerPlugin` API:
```
runtime_env:
  container:
    image: ...
    worker_path: ...
    run_options: ...
```

`ImageURIPlugin` API:
```
runtime_env:
  image_uri: ...
```

Removal of `worker_path`:
* In a followup PR, I will add the support for automatically inferring the worker path in both plugins. I will then also remove `worker_path` from the old API.

Removal of `run_options`:
* We are intentionally removing the option to specify custom `run_options` because it is too freeform; instead, we should explicitly add support for specific options as the need arises or as users submit feature requests.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>