add env var to allow dashboard http server to serve static files with symlinks (#45618)
New env var is called RAY_DASHBOARD_BUILD_FOLLOW_SYMLINKS.

This is an advanced setting that should only be used with special Ray installations
where the dashboard build files are symlinked to a different directory.
This is not recommended for most users and can pose a security risk.
Please reference the aiohttp docs here:
https://docs.aiohttp.org/en/stable/web_reference.html#aiohttp.web.UrlDispatcher.add_static

Signed-off-by: Richard Liu <ricliu@google.com>