Accelerated DAG: Support dynamic resizing of shared memory channels (#45323)
## Why are these changes needed?

In an accelerated DAG, all channels have their backing store specific
upfront. However, this size may not be large enough for all objects that
are communicated via channels, particularly since channels are reused
across multiple DAG invocations.

This PR adds support for increasing the size of a shared memory channel
backing store. Once the backing store size is increased, the size is
maintained (i.e., it is not decreased) until it needs to be increased
again.

Closes #45597

---------

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>