[Data] Explicitly specify metadata provider in `test_parquet_reader_estimate_data_size`  (#45766)
If you don't specify a metadata provider, Ray Data chooses a metadata provider to use; a specific metadata provider isn't guaranteed as part of the interface. The issue with test_parquet_reader_estimate_data_size is that it assumes that Ray Data chooses ParquetMetadataProvider. To avoid depending on this implementation details, this PR updates the test to explicitly pass the metadata provider.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>
Signed-off-by: Richard Liu <ricliu@google.com>