[Data] Use sampled fragments to estimate Parquet reader batch size (#45749)
ParquetDatasource uses fetched metadata to determine the Parquet reader batch size. If the metadata provider doesn't provide metadata, ParquetDatasource uses a default value of 10,000 rows. While this value might be reasonable for tabular data, the default value can lead to poor performance or even errors if each row is large. To avoid this issue, this PR updates the implementation to use the sampled fragments to determine the batch size.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>
Signed-off-by: Richard Liu <ricliu@google.com>