[Data] Make `read_parquet` not fetch metadata (#664)
Ray Data expects Datasource implementations to estimate the in-memory
size of data so that it can chose a good number of read tasks. To do so,
ParquetDatasource fetches metadata for all of the Parquet "fragments."
The problem is that the metadata fetching can be very slow and cause
rate limiting errors.

With the introduction of streaming generators, Ray Data doesn't need
in-memory size estimates anymore. So, this PR changes the default
metadata provider to not perform metadata fetching.

---------

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>