[runtime env] pull image during runtime env creation (#656)
[runtime env] pull image during runtime env creation

Send http request to anyscaled to pull image ahead of time during
runtime env creation, instead of during ray worker startup.

This way the image should already be pulled before raylet tries to start
the worker, and starting the worker should be very fast. This is to
avoid the image pull time triggering the timeout raylet has on starting
a worker.