Support transfer ObjectRef from JAVA application to PYTHON deployment (#45729)
## Why are these changes needed?

In the current implementation, it's not possible to transfer an ObjectRef from a Java application to a Python deployment. For more details, please refer to this issue: https://github.com/ray-project/ray/issues/45676

**Root Cause**
In the original approach, all arguments were bundled together and passed as a list parameter, like so: {1, 2, "parameter3"}. If one of these parameters was an ObjectRef, it would lead to serialization issues, preventing cross-language transfer: {1, 2, "parameter3", objectRef4}.

**Proposed Solution**
In the new approach, each argument is passed independently. The ObjectRef is implemented as PassByReference, which avoids the previous serialization issues and enables the transfer of ObjectRef: {1, 2, "parameter3", objectRef4}. 

This change enhances the interoperability between Java applications and Python deployments in Ray.