[Serve] create internal request id to track request objects (#45761)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?

User found an issue where when the request id is reused by multiple
requests around the same time, Serve's proxy will be unable to track the
request objects correctly. We should not rely on client's request id to
be unique during their lifecycle. This PR added a new
`internal_request_id` on the `RequestMetadata`, randomly generated by
the proxy, to track the request object. Client passed `request_id`
continues to be set on the `RequestMetadata` and be used for logging.
Also, updated the doc on the request id to make it more clear on what
happens when it's re-used.

## Related issue number

Closes https://github.com/ray-project/ray/issues/45723

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Gene Su <e870252314@gmail.com>
Signed-off-by: Gene Der Su <gdsu@ucdavis.edu>
Co-authored-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Signed-off-by: Richard Liu <ricliu@google.com>