[core][experimental] Driver on different node can read from accelerated DAG output (#45786)
This PR adds support for the driver reading from the accelerated DAG
output. Previously, this was not possible unless the output writer
(i.e., the last task in the DAG) was on the same node as the driver,
which may not always be the case.

To support this, the output writer needs to be able to invoke remote
functions on the driver. This is necessary so that the output writer can
create a reader ref on the driver node, and later potentially create a
larger reader ref on the driver node if the channel backing store needs
to be resized. However, remote functions cannot be invoked on the
driver.

This PR adds a DriverActor that an accelerated DAG creates when it is
initialized. This DriverActor has an empty implementation, though it
serves as a way for the output writer to invoke remote functions on the
driver node. The workflow for creating a channel that has the driver as
a reader is this:
1. Output writer creates a writer ref on its own node.
2. Output writer invokes a remote function on DriverActor to create a
reader ref on the driver node.
3. Driver gets the channel from the output writer, which already
includes the reader ref that the DriverActor created.

The workflow for resizing a channel backing store is similar:
1. Output writer detects that the channel backing store must be resized,
so it creates a new writer ref on its own node.
2. Output writer invokes a remote function on DriverActor to create a
new reader ref on the driver node.
3. Output writer sends a `_ResizeChannel` message on the channel via the
old writer-ref/reader-ref pair. The `_ResizeChannel` message includes
the new reader ref.
4. Driver gets `_ResizeChannel` message from the output writer on the
old reader ref. The driver then updates its channel to use the new
reader ref, which is included in the `_ResizeChannel` message. Note that
the DriverActor does not read from the channel, so the driver receives
this message.

---------

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>