[serve] add support for image_uri and env_vars (#674)
[serve] add support for image_uri and env_vars

Add support for using `image_uri` and `env_vars` in runtime environment
together. Whatever the user specifies in `env_vars`, the ray worker
should be started in a container with those environment variables set.

Implementation: pass the list of environment variable names to the
ray_worker_shim script. Then, since we set the environment of the
ray_worker_shim process, in the ray_worker_shim we pick up the values of
the environment variables and pass them through to the HTTP post
request. This is better than passing the key-value pairs directly as
command line arguments to the ray_worker_shim script, because it can be
tricky to handle quotation characters in the environment variable
values.