[Data] Make `ExecutionPlan.execute` return `RefBundle` instead of `BlockList` (#45852)
This PR is part of a larger effort to remove `LazyBlockList`.

Currently, `ExecutionPlan.execute()` doesn't always perform execution.
Sometimes, it skips execution and returns a `LazyBlockList`. This
`LazyBlockList` has a code path separate from the standard streaming
execution code path to load data.

This PR updates the `execute()` implementation so that it always
actually perform execution, and returns a `RefBundle` rather than a
`BlockList`.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>
Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>
Co-authored-by: Scott Lee <scottjlee@users.noreply.github.com>
Signed-off-by: Richard Liu <ricliu@google.com>