[core] Perf: Fair dispatching of the  tasks for different scheduling_class to reduce the intermediate data storage (#44733)
This PR proposes a change in the task dispatching logic within the `
LocalTaskManager` of Ray. The current implementation dispatches tasks
based on their arrival order without considering the size of the task
arguments. This can lead to suboptimal memory usage and increased task
latency, especially when larger tasks remain in the queue for longer due
to their late arrival.

By changing the dispatching order to prioritize tasks based on the size
of their arguments, we can ensure that tasks which consume large
intermediate results are processed earlier. This helps in reducing the
memory footprint as large data objects are likely to be consumed and
released sooner.

This PR introduces fairness for task scheduling. When there is a backlog
of tasks, each node will try to ensure that there is an equal number of
tasks per scheduling class running.
---------

Signed-off-by: zhilong <zhilong.chen@mail.mcgill.ca>
Signed-off-by: zhilong chen <zhilong.chen@mail.mcgill.ca>
Co-authored-by: Ruiyang Wang <56065503+rynewang@users.noreply.github.com>