[Data] Remove `in_blocks` parameter of `ExecutionPlan` (#45860)
These changes are part of a larger effort to remove LazyBlockList.

Currently, when you create an ExecutionPlan, you need to pass in a BlockList. A BlockList can either be a regular BlockList which represents in-memory data or a LazyBlockList which represents ReadTasks. The issue is that there's no way to represent lazy input data that doesn't use ReadTasks.

To address this, this PR removes the in_blocks parameter of ExecutionPlan. It follows #45852.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>
Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>
Co-authored-by: Scott Lee <scottjlee@users.noreply.github.com>