[serve] fix lossy serve config
Before: we output only the application configs, but not the system level information like http options. This PR adds system level information to the output of `serve config`.

Note that the output of `serve config` is still not identical to whatever the user submitted with `serve deploy`. We output the full system-level information, which may include defaults that the user didn't touch, e.g:
```
http_options:
  host: 127.0.0.1
```
will turn into:
```
http_options:
  host: 127.0.0.1
  port: 8000
  root_path: ''
  request_timeout_s: null
  keep_alive_timeout_s: 5
```

However the system level options, even if outputted in full, is not very spammy (unlike the application/deployment options) and the information is generally helpful as well.

Closes https://github.com/ray-project/ray/issues/40907.


Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>