[Serve] fix logging error on passing traceback object into exc_info (#46105)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?

The root cause for the logging error we saw was due to the client
passing traceback object into `exc_info` during logging. In serve logger
we do a deepcopy on the log record's `__dict__` attribute which was
unable to pickle the traceback for the deep copy. This PR removed the
deepcopy completely since we do not modify to the copied __dict__ at
all. Also added a test to ensure the regression doesn't happen.

## Related issue number

Closes https://github.com/ray-project/ray/issues/45912

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Gene Su <e870252314@gmail.com>