[Data] Add `is_lineage_serializable` to `LogicalOperator` interface (#46127)
Ray Data serializes Dataset objects so that they can be tuned as a hyperparameter for Ray Tune. To determine if a Dataset can be tuned, Ray Data explicitly checks if the input operator is a Read operator. To decouple this logic from a specific implementation and allow us to add other lineage-serializable input operators, this PR adds a is_lineage_serializable method to the interface.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>