[core][experimental] Accelerated DAG should throw an error if execution would deadlock due to NCCL ops (#45960)
NCCL ops are blocking and can deadlock if the user-specified DAG has a
cycle. To detect the deadlock,

* Convert ADAG to a graph based on the following rules:
* Add an edge from task_{bind_index} to task_{bind_index+1} on the same
actor.
* Add an edge from the writer to the reader if the channel isn't an NCCL
channel.
* Add an edge from the writer and reader of an NCCL channel to the node
that has the next bind index on the same actor as the writer.
* Perform topological sort or other algorithms to check whether the new
graph has a cycle or not. If it has any cycle, we just throw an
exception.


* Explanation: [Google
Doc](https://docs.google.com/document/d/1vwXEHHTob9rKh8GmIlaGa25EwfgVMdrG2qmwroOR48E/edit?usp=sharing)

For non-NCCL channels, the write function operates asynchronously. Take
the shared memory channel as an example, the completion of the write
function means that the data has been successfully written into the
shared memory. The completion of the write function doesn't depend on
readers. Hence, shared memory channels can't contribute to a deadlock.

## Related issue number

Closes #45594.

---------

Signed-off-by: kaihsun <kaihsun@anyscale.com>