[Doc] Turn on nitpicky mode (#46103)
## Why are these changes needed?

This PR prevents developers from breaking _internal_ sphinx references
by causing the docs build to fail if a broken reference is found.

Previously, broken references would simply be rendered as inline code
but now they will raise an exception during the build process. It will
be the responsibility of the developer who changes/adds/removes APIs or
documentation to ensure the documentation is updated accordingly.

Furthermore, single backticks (`` ` ``) will now default to just being
rendered as code; previously Sphinx considered anything in single
backticks a reference. Most developers who touch documentation are
familiar with markdown rather than RST, and use single backticks to
denote code although it is not valid RST syntax for code (RST uses
double backticks). As a result, there are many _many_ places in the
documentation that use single backticks to mean code, and it would be
impractical to expect contributors to adhere strictly to RST syntax when
everyone is used to markdown. So instead we modify the `default_role` to
eliminate this problem entirely, and treat anything in single backticks
as code.

Signed-off-by: pdmurray <peynmurray@gmail.com>