Accelerated DAGs: Performance fix for channel serialization (#45953)
A performance regression was noticed after #45545 was merged, as the
node input was changed from a single bytes object to a tuple of lists
(to support multiple arguments).

```
Before:
[unstable] compiled single-actor DAG calls per second 35571.54 +- 416.15
After Regression:
[unstable] compiled single-actor DAG calls per second 11037.4 +- 59.79
```

The default serialization engine in Ray does not need to serialize a
bytes object, but it does have to serialize other Python types,
including tuples of lists. Thus, the serialization engine needed to
perform more work with tuples of lists.

The default serialization engine took ~30 µs to serialize the tuple of
lists, which seemed expensive.

Passing a single bytes object is a common case, so this PR creates a
fast path for this case while passing everything else (multiple
arguments, nested Python objects, etc.) to the Ray serializer.

```
Disable Multi-Args (New Baseline):
[unstable] compiled single-actor DAG calls per second 28858.82 +- 270.54
Multi-Args with Fix:
[unstable] compiled single-actor DAG calls per second 27812.58 +- 536.67
```

The PR that introduced the regression was merged nearly a month ago, so
the "Before" case above is a bit stale. Thus, I disabled multi-args and
use that as the new baseline to which I compare the performance results
of this current PR.

---------

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>
Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>
Co-authored-by: Stephanie Wang <swang@cs.berkeley.edu>